kernel: Add resource limit related syscalls
This commit implements all resource limit related syscalls.

+++ Ryujinx.HLE/HOS/Kernel/Common/LimitableResource.cs
@@ -1,6 +1,6 @@
 namespace Ryujinx.HLE.HOS.Kernel.Common
 {
-    enum LimitableResource : byte
+    enum LimitableResource : uint
     {
         Memory         = 0,
         Thread         = 1,
